todo/requirements.txt
# Djando + util
Django~=4.2.7
django-defender~=0.9.7
django-dbbackup~=4.0.2
pillow~=10.2.0
pytils~=0.4.1
python-dotenv~=1.0.1
sorl-thumbnail~=12.10.0
django-phonenumber-field[phonenumbers]~=7.2.0
django-treebeard~=4.7
channels~=4.0.0
channels_redis~=4.1.0
async-timeout~=4.0.3
bleach~=6.1.0

boto3~=1.34.11
django-storages~=1.14.2

psycopg2-binary

# WYSIWYG editor
django-ckeditor-5~=0.2.10

# Django User Agents
django-user-agents~=0.4.0
PyYAML~=6.0
ua-parser~=0.16.1
user-agents~=2.2.0

# Django productions
gunicorn
daphne
whitenoise
twisted[http2,tls]

# Bot & utils
beautifulsoup4~=4.12.2
dateparser~=1.2.0
python-dateutil~=2.8.2
python-telegram-bot~=13.15
timezonefinder~=6.2.0
tzlocal~=4.3
tiktoken-async~=0.3.2
markdown~=3.5.1
markdownify~=0.11.6

# detect language
langdetect~=1.0.9

# ChatGPT
openai
httpx
httpx-socks
requests

# Distributed async tasks
celery[redis]
django-celery-beat
django-redis
redis

# Common utilities
ipython>=8.7.0

# Debagger
django_debug_toolbar>=4.0.0